chore: load reading data without resources
